+++ CHANGELOG.md
@@ -6,6 +6,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] ##
+
+### Added ###
+
+* End message, saying how many tests passed and how many failed
+
+[Unreleased]: https://github.com/SpEZiiL/utest-script/compare/v0.1.0...develop
+
 ## [v0.1.0] - 2019-06-10 ##
 
 ### Added ###